[Added] Named IBL types in C++ and shader side [Added] Extra API to set IBL properties individually

+++ filament/src/PerViewUniforms.cpp
@@ -216,7 +216,7 @@ void PerViewUniforms::prepareDirectionalLight(
 
 void PerViewUniforms::prepareIblLight(const IblOptions& options) noexcept {
     auto& s = mPerViewUb.edit();
-    s.iblTechnique = options.iblTechnique;
+    s.iblTechnique = static_cast<uint32_t>(options.iblTechnique);
     s.iblCenter = options.iblCenter;
     s.iblHalfExtents = options.iblHalfExtents;
 }
